Separated robot logic, created separate ruby test file
